﻿https://extract.atlassian.net/browse/ISSUE-964 Fixed VS designer crashed when attempting to delete non-empty DataEntryGroupBox controls.
https://extract.atlassian.net/browse/ISSUE-12353
Removed assertion that would fire in the VS designer that incorrectly warned of combo box table cells that supposedly didn’t have appropriate validation lists. Since this assertion hadn’t been warning when it should have been either (and it’s not that important of an assertion in my opinion) I removed the assertion.

[git-vault-id] Extract$/Engineering@29820/71927