﻿https://extract.atlassian.net/browse/ISSUE-12447 Added new Specified File(s) condition to the FAM DB Admin’s select files dialog. Allows one or more specific filenames to be selected from the DB (either manually specified or via a specified text file list). Also supports wildcards to be able to select all files in a particular folder, for instance.
https://extract.atlassian.net/browse/ISSUE-12429
Added new File set condition to the FAM DB Admin’s select files dialog. This allows a defined FileProcessingDB file set to be used to select files. Currently file sets are not persisted to the DB (exist only for the lifetime of a FileProcessingDB instance). The only way they can currently be created is programmatically, though the special button to be added for UW’s duplicate document solution will make use of this. The File set condition is only available when at least one file set has been defined.

As part of this, added new methods to FileProcessingDB to allow creation and retrieval of file sets.

NOTE: File set functionality cannot currently be tested since there are no places that file sets are currently create in code.

https://extract.atlassian.net/browse/ISSUE-12446
Fixed DB schema update from < version 116 to >=117. This had broken in relation to the added “Open file location” feature added for the FFI.

Reviewed by:
William (3/2)

[git-vault-id] Extract$/Engineering@30380/72499