﻿https://extract.atlassian.net/browse/ISSUE-12445 Fixed bug in $Replace() path tag that would cause incorrect replacements when there is more than one instance of the search term to replace.
https://extract.atlassian.net/browse/ISSUE-12410
Fixed issue where extra parentheses in a string being expanded with path tags/functions would cause errors expanding the string.
Added an automated test for this situation.

Fixed NUnit configuration so that the debug profile expects the engineering tree to be on the C: drive instead of the D: drive.

Review pending…

[git-vault-id] Extract$/Engineering@30381/72500