﻿https://extract.atlassian.net/browse/ISSUE-12439 Fixed problem where only the last referenced attribute in a query mode with a selection mode of distinct would cause the query to be re-evaluated. (so that changes all other referenced attributes would not trigger queries to fire)
https://extract.atlassian.net/browse/ISSUE-12453
It was possible for incorrect results to be returned from queries when operations that cause multiple attribute manipulations (such as initializing a new row in a table) caused queries to be executed numerous times through the operation.  I put in place a system that delays query execution until such operations are complete so that the queries are executed against the final result of the operation (and not all its intermediate states as well).

Query responsiveness should be improved as a result of these changes… though don’t know if it will add up to a noticeable difference.

Reviewed by:
Nat (0/6)
Nat (1/2)

[git-vault-id] Extract$/Engineering@30464/82587