﻿$/Engineering/ReusableComponents/COMComponents/UCLIDFileProcessing/Code
https://extract.atlassian.net/browse/ISSUE-12388
Priority works better with workitems.
1.	Process work items in local queue
2.	When work item queue is empty load work items with priority >= priority of next file to process
3.	If there are no work items from 2 then next file gets processed

Note:  if a higher priority file is added, the FAM will process all of the files that are in the local queue before the higher priority file will be processed

Example: FAM setup with 8 threads and “Files to retrieve setting” = 10
1.	Add 1000 page document at normal priority while a FAM is processing.
2.	Add 50 page document at Above Normal priority while the 1000 page document is processing
a.	All the work items that are in the local queue will be processed
b.	When more work items are requested from the DB, it requests those with the highest file priority that is Above normal or higher
c.	Since there are no work items with that priority, the next file starts processing
d.	File gets split into work items and they then get loaded.

https://extract.atlassian.net/browse/ISSUE-12374
WorkItems are retrieved from the database in groups based on the “Files to retrieve setting”

William

[git-vault-id] Extract$/Engineering@30715/82841