﻿Changed default stapled output path for the UW duplicate documents FFI to: $DirOf($DirOf(<SourceDocName>))\Lab\Stapled_<FirstName>_<LastName>_$Now().tif
Added validation when outputting stapled documents that if the configured output path begins with:
$DirOf($DirOf(<SourceDocName>)), that this portion of the path evaluates to the same value for all files being stapled (validates that the grandparent folder of all source files is the same).

https://extract.atlassian.net/browse/ISSUE-12527
Fixed issue where UW dup docs FFI could crash if processing was stopped via the FAM while a cell was in edit mode in the FFI. This appears that it may trace back to a deficiency of sorts in the .Net framework—it appears in general that if a DataGridView is disposed while in edit mode it will throw an exception (which is unlikely to be handled). Therefore, I have added a similar fix for places in IDS and DE verification to prevent against this from ever happening there as well.

Reviewed by:
Nat (0/0)

[git-vault-id] Extract$/Engineering@30980/83108