﻿https://extract.atlassian.net/browse/ISSUE-12528 Fixed issue that was preventing processing files from being set back to pending in some circumstances such as when processing stops due to a processing schedule.
This was introduced as part of the code for IFileRequestHandler (ISSUE-12449). Since I believe a contributing factor in the introduction of this bug was that the order of the parameters bQueueChangeIfProcessing and bAllowQueuedStatusOverride was opposite in setFileActionState as it was in setStatusForFile, I have also reversed the order of those two parameters in setFileActionState so as to prevent any future mix-up of these parameters.

Reviewed by:
William (0/1)

[git-vault-id] Extract$/Engineering@30987/83115