﻿$/Engineering/ProductDevelopment/Common
Fixed misspelling of FixMachineConfig.vbs file

William

From: William Parr [mailto:william_parr@extractsystems.com]
Sent: Friday, November 7, 2014 2:59 PM
To: ag@extractsystems.com; Nathaniel_heyer@extractsystems.com; steve_kurth@extractsystems.com; wayne_lenius@extractsystems.com; william@extractsystems.com
Subject: Check-in

$/Engineering/ProductDevelopment/Common
$/Engineering/ProductDevelopment/AttributeFinder/Build

https://extract.atlassian.net/browse/ISSUE-12519
Added FixMachineConfig.vbs script to install that will detect the invalid <DbProviderFactories/> tag if there is a <DbProviderFactories>
If found the script will check
1.	if it is running on OS > XP
a.	if not it proceeds as if it has Admin rights and will fail with a message if it does not
2.	Checks if it has Admin rights
a.	If does not have admin rights and running on os >XP, it prompts for elevation
3.	Checks for previous backup copy – fails with message if backup exists
4.	Updates the machine.config file
5.	Displays message that the change was made
If the problem was not found, a message will be displayed telling the user the problem wasn’t found.

Reviewed by Nat(0/0)

William

[git-vault-id] Extract$/Engineering@31183/83311