﻿$/Engineering/ProductDevelopment/Common/OSSI/SQLServer $/Engineering/ProductDevelopment/Common/OSSI/SQLServerMgr
https://extract.atlassian.net/browse/ISSUE-12584
Extracted the files from the SQL Server 2014 install and the SQL Management Studio 2014 install and modified the OSSI.ini file for each install to call the extracted setup file.  The OSSI files have been updated in the \\engsvr\internal\ProductReleases\SharedInstalls\SQLServerExpress2014 and \\engsvr\internal\ProductReleases\SharedInstalls\SQLServerExpress2014Mgr

Using the extracted installs increases the combined size of these folders by about 1.13 GB. The size of the LabDE install is the largest at 4.25GB which is getting close to the limit of what will fit on a dvd.

William

[git-vault-id] Extract$/Engineering@31356/83484