﻿$/Engineering/ProductDevelopment/AttributeFinder/AFCore/Code $/Engineering/ReusableComponents/COMComponents/UCLIDCOMUtils/Core/Code
https://extract.atlassian.net/browse/ISSUE-12293.
Fixed problem of copy and paste of some rules objects(preprocessors, output handlers, and finding rules) keeping the same GUID if they were identifiable objects.  The problem was that when some of these objects are pasted an exact copy would be made by streaming to and from the clipboard.  It is better to Clone the object when pasting.  This change was implemented for copy and pasting File suppliers and Tasks in the FAM also.

Reviewed by Nat(0/0)

William

[git-vault-id] Extract$/Engineering@32009/84146