﻿https://extract.atlassian.net/browse/ISSUE-12804 Added framework for ExtractSettings classes to expand path tags values independent of whether the settings are stored in a config file, registry, or FAM DB. For the ConfigFileSettings class in particular, all settings outside of <objectSettings> will now by default have path tags and functions expanded, while values in <objectSettings> will be interpreted as literal XML by default. All values in DE config files will now support the attributes “XML” and “expandTags” for  settings/properties  (Both bools such as XML=”True” or expandTags=”False”).
 If XML is true (default in <objectSettings>), the value will be treated as literal XML, otherwise any XML escape sequences (such as “&lt;”) will be un-escaped.
If expandTags is true (default in <applicationSettings> and <databaseConnections>), any path tags/functions will be expanded.

NOTE: XML and expandTags cannot both be set to true (since angle brackets in path tags will cause malformed XML).
NOTE2: Path tag expansion will happen as the settings are first loaded and not be thereafter re-evaluated. Thus, tags that have values that change (such as <SourceDocName>) cannot be used. If a need arises in the future, specific settings could be expanded at the point of usage and be used with expandTags=”False” in the config file.

Examples:
In applicationSettings/databaseConnections:
<DatabaseConnectionString expandTags=”True”>Data Source=&lt;DatabaseServer&gt;;Initial Catalog=&lt;DatabaseName&gt;;Integrated Security=True</DatabaseConnectionString>
In objectSettings:
<Property name="AutoUpdateQuery" XML="False" expandTags="True">&lt;FPSFileDir&gt;</Property>

Reviewed by:
William (2/0)

[git-vault-id] Extract$/Engineering@32384/84537