﻿https://extract.atlassian.net/browse/ISSUE-12806 Added ability for LabDE DEP to be able to check and prompt about order results that have already been submitted and to be able to associated orders that are being committed with the document they are from.
https://extract.atlassian.net/browse/ISSUE-12932
Added framework to allow for the DataEntry framework to clear all data that had been cached from databases and used it to add a refresh button in the demo DEP. This refreshes the data both the custom LabDE controls as well as the data query system.

https://extract.atlassian.net/browse/ISSUE-12933
Added new BasicMenuOptionsOnly property for custom FFI columns and selection panes to allow them to hide database and custom FAM file handler menu options. The dup docs and order picker contexts take advantage of this.

https://extract.atlassian.net/browse/ISSUE-12803
Cleaned up all known remaining loose ends with the order picker:
- Fixed an error that would result with alternating between opening the duplicate documents FFI and the order-picker FFI.
- Fixed an error that would sometimes occur if the order picker FFI was close very quickly after opening.
- Integrated the data commit/cancel and OK/Cancel buttons for the order picker FFI with the existing custom data commit framework created for custom columns.
- Fixed issue where order selection grid did not properly resize when the pane it was in was resized.
- Added the patient name to the order summary information displayed in the order picker / order number tooltip.

Reviewed by:
William (1/1)

[git-vault-id] Extract$/Engineering@32861/95011