﻿$/Engineering/ReusableComponents/COMComponents/UCLIDFileProcessing/Utils/ReportViewer/Core/Code $/Engineering/RC.Net/UtilityApplications/SQLCompactUtils/SQLCDBEditor/Core/Code $/Engineering/RC.Net/UtilityApplications/PaginationUtility/Core/Code $/Engineering/RC.Net/Utilities/Forms/Core/Code $/Engineering/RC.Net/Utilities/Core/Code $/Engineering/RC.Net/DataEntry/Utilities/DataEntryApplication/Core/Code
https://extract.atlassian.net/browse/ISSUE-12830
Modified code so that the user.config file will get checked for corruption early in the application and rechecked before saves of data and makes a backup copy that can be used if corruption is found. If the backup does not exist the corrupt user.config file will be deleted.  There may still be cases that are not completely covered because the loading of settings from the User.Config file is something that .Net does “automatically” and this code checks the user.config file before the “automatic” loading happens.

Reviewed by Steve(4/2)

William

[git-vault-id] Extract$/Engineering@33008/95159