﻿ https://extract.atlassian.net/browse/ISSUE-13003 Fixed an exception that was raised by the order picker column when there was more than one order row on a document that could not be mapped to any orders in the LabDEOrder table and that sometimes also occurred after deleting rows.
https://extract.atlassian.net/browse/ISSUE-13006
Added AutoPopulate property to the OrderPickerTableColumn which causes the order number to be automatically populated if is the one and only row that would be displayed in the order picker FFI and match any specified AutoSelectionFilter.

https://extract.atlassian.net/browse/ISSUE-13013
The OrderPickerSelectionPane will no longer appear in the Visual Studio designer toolbox when Extract.DataEntry.LabDE.dll is referenced.

Reviewed by:
William (0/0)

[git-vault-id] Extract$/Engineering@33177/95341