﻿https://extract.atlassian.net/browse/ISSUE-12987 https://extract.atlassian.net/browse/ISSUE-13001
Fixed a memory leak introduced by the plumbing behind the “Refresh database data” button now available for DataEntry DEPs. The fix virtually eliminated the memory leak discovered that had affected an FPS to add files and metadata to a database, but a DataEntry preloader test (ISSUE-12987) still shows somewhat of a memory leak, albeit > 8x less than what it had been and probably passable for a 10.1 release if need be.

Reviewed by:
William (0/0)

[git-vault-id] Extract$/Engineering@33224/95388