﻿$/Engineering/RC.Net/UtilityApplications/PaginationUtility/Core/Code
The PaginationUtility.csproj file was still specifying the vault server as White instead of Engsvr in the SccAuxPath.  I believe the setting in the solution file overrides this so it wasn’t causing a problem, but if the project file is opened directly it would try to connect to White.

William

[git-vault-id] Extract$/Engineering@33384/95548