﻿$/Engineering/ProductDevelopment/Common/Powershell $/Engineering/ProductDevelopment/AttributeFinder/Build
Powershell script was checked into the wrong folder and the full path needed to be used for the call to the script. Also modified the order of the calls so that the email for the start of a build will be just after the version is updated but before it is commited.

William

_____________________________________________
From: William Parr [mailto:William@extractsystems.com]
Sent: Tuesday, June 16, 2015 9:27 AM
To: ag@extractsystems.com; Nathaniel_heyer@extractsystems.com; steve_kurth@extractsystems.com; wayne_lenius@extractsystems.com; william@extractsystems.com
Subject: check-in

$/Engineering/ProductDevelopment/AttributeFinder/Build
$/Engineering/ProductDevelopment/Common/Powershell

https://extract.atlassian.net/browse/ISSUE-12913
Modified build batch files to send email when a build is started and when it finishes indicating if it was successful or it failed.

William

[git-vault-id] Extract$/Engineering@33662/95835