﻿$/Engineering/ProductDevelopment/AttributeFinder/Build $/Engineering/ProductDevelopment/AttributeFinder/IndustrySpecific/Redaction/Installation/IDShield $/Engineering/ProductDevelopment/AttributeFinder/IndustrySpecific/Redaction/Installation/IDShieldInstall $/Engineering/ProductDevelopment/AttributeFinder/Installation/FlexInstall $/Engineering/ProductDevelopment/AttributeFinder/Installation/UCLID FlexIndex SDK $/Engineering/ProductDevelopment/Common/InstallScripts $/Engineering/ProductDevelopment/Common $/Engineering/ProductDevelopment/DataEntry/LabDE/Installation/LabDE $/Engineering/ProductDevelopment/DataEntry/LabDE/Installation/LabDE/Script Files $/Engineering/ProductDevelopment/DataEntry/LabDE/Installation/LabDEInstall $/Engineering/RC.Net/FileActionManager/FileSuppliers/Core/Code $/Engineering/RC.Net/Utilities/Core/Code
https://extract.atlassian.net/browse/ISSUE-12692
Removed the code in the install that registers the shell extentions and changed the .nl files that are created to no longer include the 2 dlls: Extract.Utilities.ShellExtensions.dll and LogicNP.EZShellExtensions.dll

Added 2 batch files: RegisterShellExtension.bat and UnRegisterShellExtension.bat that can be run to register and unregister the 2 dlls.

Modified code for the ContextMenuFileSupplier so that it will not show up in the list of suppliers if either Extract.Utilities.ShellExtensions.dll and LogicNP.EZShellExtensions.dll are not registered. If it exists in a FPS file that was previously created if that FPS file is ran an exception will be thrown that indicates that those dlls are not registered and it will suggest running

Reviewed by Steve(0/0)

https://extract.atlassian.net/browse/ISSUE-13112
Modified the Demoshield text to indicate that our products only require .NET 4.0

https://extract.atlassian.net/browse/ISSUE-13111
Removed .NET 3.5 as a prerequisite for our products (LabDE, ID Shield, FLEX Index)

William

[git-vault-id] Extract$/Engineering@33938/96111