﻿$/Engineering/ProductDevelopment/AttributeFinder/Build
This did not work so I have undid that change and modified the calls to nmake to save the stderr results to a file and then check the file for errors as a way to detect a failed build.  This may change the contents of our log file – the nmake errors may not be there.

William

_____________________________________________
From: William Parr [mailto:William@extractsystems.com]
Sent: Friday, June 26, 2015 2:23 PM
To: ag@extractsystems.com; Nathaniel_heyer@extractsystems.com; steve_kurth@extractsystems.com; wayne_lenius@extractsystems.com; william@extractsystems.com
Subject: check-in

$/Engineering/ProductDevelopment/AttributeFinder/Build

https://extract.atlassian.net/browse/ISSUE-12913
Modified settings on call to nmake to have it return errorcode of 1 if it fails.

William

[git-vault-id] Extract$/Engineering@33994/96167