﻿$/Engineering/RC.Net/FileActionManager/Utilities/ESFAMService/Core/Code $/Engineering/ReusableComponents/COMComponents/UCLIDFileProcessing/Utils/FAMDBAdmin/Code
https://extract.atlassian.net/browse/ISSUE-13049
Modified FAMDBAdmin to disable the Metadata Fields menu if the database has an older schema

https://extract.atlassian.net/browse/ISSUE-2039
Modified behavior so that if a fps file used by a service is configured to  “Stop processing files once the queue is empty” set that exceptions will be logged every time it is started that indicate it is not a recommended configuration. An exception will also be logged anytime a fps file stops processing or process is killed that will not be restarted.

Reviewed by Steve (0/2)

William

[git-vault-id] Extract$/Engineering@34047/96220