﻿https://extract.atlassian.net/browse/ISSUE-13185 Created a FAMMetadataManager class that provides  programmatic method for DEPs to update FAM DB metadata on save (instead of using RunFPS to call a FPS to run a ruleset with a DataQuery).
https://extract.atlassian.net/browse/ISSUE-13196
Fixed a but that prevented both the XML and expandTags attributes for .Net config files from both being explicitly specified on the same value.

Added new automated tests to Extract.Utilities.Test for the XML and expandTags attributes.

Updated Demo_LabDE to use the FAMMetadataManager class and tweaked the setting of the StapledDocumentOutput directory so it works if the CustomTags.sdf are modified so that the demo runs somewhere other than C:\Demo_LabDE.

Reviewed by:
William (0/1)
David (0/0)

[git-vault-id] Extract$/Engineering@34374/96552