﻿https://extract.atlassian.net/browse/ISSUE-12789 Factored all UPI and UPIID columns out of the database and replaced them with a foreign key to the FAMSession table. Because the schema now requires it, the option for whether to log to the FAMSession table has been removed.
https://extract.atlassian.net/browse/ISSUE-11381
The FAM status bar’s processing value now reflect the number of files processing in the database rather than the number of files processing in the current FAM. (User can now look at the processing files count on the processing log tab if they want to see the number processing in the current FAM).

https://extract.atlassian.net/browse/ISSUE-13215
The FAM status bar now updates to reflect the counts in a database after connecting to the database.

Reviewed by:
William (2/2)

[git-vault-id] Extract$/Engineering@34633/96812