﻿$/Engineering/ProductDevelopment/AttributeFinder/AFCore/Code $/Engineering/ProductDevelopment/AttributeFinder/IndustrySpecific/Redaction/RedactionCustomComponents/Code $/Engineering/ProductDevelopment/AttributeFinder/Utils/EAVGenerator/Code $/Engineering/RC.Net/AttributeFinder/Core/Code $/Engineering/RC.Net/Redaction/Core/Code $/Engineering/RC.Net/Redaction/Verification/Core/Code $/Engineering/ReusableComponents/COMComponents/UCLIDCOMUtils/Core/Code
https://extract.atlassian.net/browse/ISSUE-13202
Modified Attributes to have an instanceGUID – added a new interface ICloneIdentifiableObject with 2 methods- CloneIdentifiableObject and CopyFromIdentifiableObject to duplicate the functionality of the ICopyableObject interface except that instead of creating a new InstanceGUID, the InstanceGUID is copied.

I did find a situation that would lead to duplicated ID’s. In the IDShield Verification if you turn off a previously saved redaction and then turn it back on in another session, the top level attribute and the _IDAndRevision attributes will be the same as the _OldRevisions version of the attribute. This is not straight forward to resolve. I will add a jira entry for it

Reviewed by Steve(1/1)

William

[git-vault-id] Extract$/Engineering@34680/96859