﻿$/Engineering/ProductDevelopment/AttributeFinder/SilentInstalls
https://extract.atlassian.net/browse/ISSUE-13240
Fixed LabDEuninst.iss file to have the correct response to dialogs for Uninstalling Crystal Reports and SQLCE so that the uninstall dialog is not displayed when doing a silent install if LabDE is installed when the silent install is ran.

This problem is in both the 10.1 and 10.2 silent installs.

William

[git-vault-id] Extract$/Engineering@34799/96978