﻿William_Parr:  Adds a new DBInfo value DatabaseID that is an encrypted string of the format <DatabaseGUID>,<DatabaseServer>,<DatabaseName>,<DatabaseCreationDate>,<DatabaseRestoreDate> This value should always be kept when clearing the database even if the check box to retain values is unchecked.
Adds the EncryptCounter and EncryptCounterChangeValue tables - this are also kept regardless of the check box to retain values.

I also made a change to the method that drops foreign keys so that if either of the tables used in the FK are in the list of tables being dropped the FK will be dropped.
[git-vault-id] Extract$/Engineering@34963/97142