﻿$/Engineering/ReusableComponents/COMComponents/UCLIDFileProcessing/Code $/Engineering/ReusableComponents/COMComponents/UCLIDFileProcessing/FAMUtils/Code
https://extract.atlassian.net/browse/ISSUE-13247
Added 2 tables to support counters – SecureCounter and SecureCounterValueChange tables – the tables will not be deleted when a clearing the database. As part of this change the code the drops foreign keys to drop FK if either one of the tables that are part of the FK are going to be dropped.

https://extract.atlassian.net/browse/ISSUE-13246
Adds a new DBInfo value DatabaseID that is an encrypted string of the format
<DatabaseGUID>,<DatabaseServer>,<DatabaseName>,<DatabaseCreationDate>,<DatabaseRestoreDate>
This value should always be kept when clearing the database even if the check box to retain values is unchecked.

Reviewed By Steve(2/2)

William

[git-vault-id] Extract$/Engineering@34964/97143