﻿https://extract.atlassian.net/browse/ISSUE-13273
Add new option to OrderMapper that allows filled requirements and mandatory requirements to be strictly or loosely enforced.
Add some short-circuit logic for cases where filled, mandatory, or outstanding order requirements are set to be enforced but where no requirements are defined in the database/input.
Fix bug with round-trip logic that was revealed when implementing the above short-circuit logic.
Add automated test cases.

Reviewed by:
Steve (2/1)
[git-vault-id] Extract$/Engineering@35771/97958