﻿$/Engineering/ReusableComponents/COMComponents/UCLIDFileProcessing/Code $/Engineering/RC.Net/FileActionManager/Database/Core/Code
https://extract.atlassian.net/browse/ISSUE-13459
Modified code so that if the DatabaseID value is empty (but the record exists in DBInfo) and there are no counters an update code can be generated so that new counters can be created. The previous change would throw an exception.

https://extract.atlassian.net/browse/ISSUE-13470
Change the send alert code to log instead of throw an exception

https://extract.atlassian.net/browse/ISSUE-13469
Added a warning that the SMTP settings need to be updated before being able to enable alerts

Reviewed By
David(0/0)
Steve(0/2)

William

[git-vault-id] Extract$/Engineering@35988/98176