﻿ISSUE-13401 Create <Add new> item for creating new attribute sets dynamically
Made one change to attributeDbMgr.cpp - now SQL Encodes the attribute set name input to CAttributeDBMgr::CreateNewAttributeSetName(). This enables single quote characters in the name.
The rest was very straight-forward. "Add new..." is only allowed when the dialog mode is set to "Store", not "Retrieve", as the latter case doesn't make sense.

[git-vault-id] Extract$/Engineering@36356/118545