﻿https://extract.atlassian.net/browse/ISSUE-13550 Added new property EnableRowAutoPopulation to the DataEntryTable class that defaults to true, but allows auto-population of data based on smart hints to be disabled.
https://extract.atlassian.net/browse/ISSUE-13549
Tweaked previous case rule that prevented auto-population of rows that  have any data so that it instead prevents auto-population of a row that has been modified in any way since being inserted. This allows for the case that a column can have a default value that is set when a new row is inserted (such as the status column of the result components table in Demo_LabDE).

Reviewed by:
David (0/1)

[git-vault-id] Extract$/Engineering@36416/118605