﻿https://extract.atlassian.net/browse/ISSUE-13481 Setting an alert level no longer causes the alert frequency to appear as set the next time the dialog is opened when it is not set.
https://extract.atlassian.net/browse/ISSUE-13466
Added prompts to save changes to rule execution counter alert changes on cancel and changed behavior so that alert level changes aren’t lost when refreshing counter values.

https://extract.atlassian.net/browse/ISSUE-13507
Modified SetDBInfoSetting to have a flag that indicates whether the change should be recorded in DBInfoChangeHistory and set it to false in the legacy cases it was used (prod-specific DB managers on create/clear/update) as the lack of a FAMUser was causing some operations to fail by recording history. Making the value of this flag be more nuanced based on circumstance in legacy locations might be the technically more appropriate thing to do, but no one has ever brought this up in the many years the prod-specific settings history has been missing and I’d rather error on the side of caution.

Also modified rule execution management dialog so that counters always appear in order of counter ID.

Reviewed by:
Nat (0/4)

[git-vault-id] Extract$/Engineering@36470/118659