﻿https://extract.atlassian.net/browse/ISSUE-13605 Abstracted into DbMethods code from SQLCDBEditor to allow for formatting DataGridViews from query results. Allows for reusable formatting of DataGridViews without requiring that the grid be managed by the SQLCDBEditor application.
Also changed the UIReplacementPlugin property of IDatabaseSchemaManager to UIReplacementPlugins to allow multiple plugins to replace the editor’s normal UI instead of just one. For the time being, I have not commented in the code to allow the LabDE self-service plugins to replace the normal UI to not create inconvenience when OrderMappingDBs need to be edited. In a follow-up check-in I will at least provide a backdoor to open the new UI plugins.

Reviewed by:
Nat (4/6)

[git-vault-id] Extract$/Engineering@36847/119036