﻿https://extract.atlassian.net/browse/CUST-1853 UW Clinics DEP now automatically sets the abnormal flag based on the value and range.
https://extract.atlassian.net/browse/CUST-1869
Apostrophes in the patient name fields will no longer cause problems with the validation queries.

https://extract.atlassian.net/browse/CUST-1855
Fixed an error that occurred with a query in the UWT DEP for 10.3, though I’m concerned about a possible underlying issue with the OrderCode field that I can’t reproduce or track down.

NOTE: I was unable to check-in the DEP dll for UWT since it is currently exclusively locked via the services account.

[git-vault-id] Extract$/Engineering@37744/139954