﻿https://extract.atlassian.net/browse/ISSUE-13802 Fixed exception that occurred when switching between the data and pagination tab with no file loaded.
https://extract.atlassian.net/browse/ISSUE-13803
Fixed issues with the windows title bar text if on the pagination tab when a new document comes into the queue.

https://extract.atlassian.net/browse/ISSUE-13901
Fixed memory leak in stand-alone pagination task.

https://extract.atlassian.net/browse/ISSUE-13903
FAM now correctly advances to next file if all pages of a source document are deleted.

https://extract.atlassian.net/browse/ISSUE-13908
Fixed issue where having a copied page on the clipboard from a previous document could cause that document to be put back in the pagination tab when reverting pagination on a later document.

https://extract.atlassian.net/browse/ISSUE-13909
Pagination tab is no longer cleared on close/skip before user has had opportunity to cancel.

Also:
- Fixed issue where page selection was not defaulting correctly on the pagination panel
- Fixed loophole that I think has in a couple cases been the reason I have gotten stuck on the pagination tab and not been able to get back to the data tab.
- Fixed case where an image could get left open after applying pagination.
- Removed external file drop capability which is not appropriate for FAM-based processing.

Review pending…

[git-vault-id] Extract$/Engineering@37863/140073