Document.GetResultSet() now works, including returning complete DocumentAttribute - so mapping of FAM Attribute to DocumentAttribute is complete, and the DocumentAttributeSet now conforms to the 2-16-2017 meeting conclusions. Also refactored so that options, fileprocessingDB, and AttributeMgrDB are all set up in Utils
added DatabaseController (for testing) and the GeneratedApi project + files

Fixed an issue in the AttributeMapper where RasterZones were not being enumerated. Also made a few minor mods.

Modifications to support testing, including add of test controller and regeneration of client API

Converted FileAPI_VS2017 project to DocumentAPI and moved it from Web_API_POC to WebApi - and did the same with the GeneratedAPI project.

Completed unit test for Document.GetResultSet(), and some cleanup in the DocumentAPI project.

cleanup on Workflow.cs