Added code that should prevent inappropriate assertion related to changing workflow while processing. https://extract.atlassian.net/browse/ISSUE-14766
Fixed issue with moving files between workflows that are unattempted in all actions.
Added a unit test for this scenario
https://extract.atlassian.net/browse/ISSUE-14800

Fixed reported count of files moved between workflows to correctly indicate the count of files moved as opposed to the total number of files in the selected fileset (some of which may already exist in the target workflow).