Made a fix to allow hints to work in DE verification. (had been broken since the fix for ISSUE-14328 in 10.5) https://extract.atlassian.net/browse/ISSUE-14901
Created Split multi-page document file processing task that will split a multi-page document into single page documents and create uss and voa files for the split document based on the source document's uss and voas. The new files will be added to the DB and associated with the parent via the Pagination table.
https://extract.atlassian.net/browse/ISSUE-14942

Reviewed by Nat