Added ability to show highlights on the Dashboard Detail form https://extract.atlassian.net/browse/ISSUE-15357
In order for the highlights to be shown for the selected item, query that is configured for the data in the Configure Extract Settings for a dashboard grid requires the following field names
ExpectedSetForFileID - ID from the AttributeForFile table for the record that contains the VOA for the expected set
ExpectedGuid - GUID for the Expected Attribute to be displayed ( this guid can be in a VOA for multiple attributes because attribute history that can be stored int the VOA)
ExpectedAttributePath - Full path using Expected Attribute names and sub attribute names to get to the attribute that should have a Guid matchin ExpectedGuid this path should have names separated by /

FoundSetForFileID - ID from the AttributeForFile table for the record that contains the VOA for the found set
FoundGuid - GUID for the Found Attribute to be displayed ( this guid can be in a VOA for multiple attributes because attribute history that can be stored int the VOA)
FoundAttributePath - Full path using Found Attribute names and sub attribute names to get to the attribute that should have a Guid matching FoundGuid this path should have names separated by /

The Found highlights are displayed as Coral the Expected attributes are Aqua.  Where the highlights overlap the color will be green.

Added Debug builds of DashboardCreator and DashboardViewer