Fixed issue with swipes not updating document summary text properly. https://extract.atlassian.net/browse/ISSUE-15472
Removed tweak to visual style that was intended to allow active field indication for combo boxes, but is suspected of a crash seen @ Duke.
https://extract.atlassian.net/browse/ISSUE-15445

"Restore as originally loaded" now restores original viewed status and checkbox state that indicates document selected to apply.
https://extract.atlassian.net/browse/ISSUE-15474

Updated pagination panel in data entry verification with pagination to have same gray background as is used in the pagination task.