Training Coordinator: Database timeout when running pagination collector https://extract.atlassian.net/browse/ISSUE-15662
This moves most of the work for LearningMachine data collection outside the transaction. NER annotator is still doing too much in the transaction but this is usually pretty fast...
I have put in code to retry with small batch sizes on failure and to skip files if batch size of 1 fails. It is crude but should prevent stalls