Utility and fixes for splitting and filtering RTF file packages into individual text and uss files https://extract.atlassian.net/browse/CUST-3432
Still to be done is making a task that creates USS files and maybe writes to pagination table. This will most-likely replace the exe so I have not written unit tests for the exe nor added it to the install.
Necessary work to do the replacement will be done as part of https://extract.atlassian.net/browse/ISSUE-15794