Changed the VS2015 redistributes included to ones in Common Files folder https://extract.atlassian.net/browse/ISSUE-15941
There was one merge module that was included but the one that said it had api-ms-crt-runtime-|1-1-0.dll was included.
I believe this were manually added from a different location and were not previously used by our software so we had no problems.
We compile our code against the vs2010 library