Added ability for pagination verification to indicate documents that have qualified for automatic processing by the auto-paginate task. https://extract.atlassian.net/browse/ISSUE-16655
Added ability for pagination verification to indicate documents that have validation warnings via the document header bar.
https://extract.atlassian.net/browse/ISSUE-16656

Also updated the auto-paginated task so that when not configured to output documents, source documents will be routed to the source document action for files not fully paginated. This allows the workflow to be changed from a preview of auto-pagination to full auto-pagination with a single change to the auto-pagination settings. Also, source action for fully paginated files is no longer required.