Changed DatabaseID generated on SQL Server clusters to retrieve the name of the Cluster instead of the SQL Server connected to https://extract.atlassian.net/browse/ISSUE-15822
There is also a change to not check the Creation date and restore date if a cluster is being used because those dates are different for each replica/instance that is part of the cluster.

The Cluster name is retrieved using a SQL stored procedure. The retrieval of the Cluster name does require a user with "View Server State" permission