FAM: Fix edit-action behavior when specified with path tags
Change so that the previous action text is shown when editing, even if
the action is not part of the currently-connected-to database. This is
primarily to make it easier to edit the action if it is specified with
path tags/functions.

NOTE: There was an essentially unused action ID field that has now been
deleted.

https://extract.atlassian.net/browse/ISSUE-18918