ErrorsAndAlerts: Environment Information View & ViewModel Bindings
https://extract.atlassian.net/browse/ISSUE-18937

Changed the view to match new data being used for environment information. Basic variables and bindings set up to appropriate view model.

Changes made to EventsOverallViewModel to allow it to open a window with this view have broken 2 unit tests that test constructors, since the view is unmocked. This was considered to be fine, since the tests are covering currently unimplemented code, and will be fixed at implementation.