Address inefficiencies in Pagination verification (#124)
Address inefficiencies in Pagination verification:
https://extract.atlassian.net/browse/ISSUE-19112

Major: AttributeStatusInfo.ReleaseAttributes has a recursive loop that was taking a majority of the time required to close a document. Replacing the recursion with a simple loop removed most of the processing time

Major: PaginationPanel.AutoRotateImages was re-opening and parsing the uss file for each logical document in the source document for the purpose of determining image page orientation.

Minor: Despite the fix for ISSUE-18904 _SharedData attribute is persisting in VOAs when using SupportsNoUILoad = True

Minor: There’s a fair amount of processing that occurs in Attribute::StowDataObject and the translation of the stowed data to output data in Attribute::Save that could be avoided by stowing directly to a byte array rather than back-and-forth to stringized byte streams.